requirements.txt
accounts==0.2.4
alembic==1.7.7
amqp==2.6.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
atpublic==3.0.1
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.0.0
celery==4.4.7
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cssselect==1.1.0
cssutils==2.4.0
cycler==0.11.0
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.4
Django==3.2.13
django-allauth==0.50.0
django-appconf==1.0.5
django-autoslug==1.9.8
django-braces==1.15.0
django-cities==0.6
django-colorfield==0.6.3
django-countries==7.3.2
django-countries-plus==1.3.2
django-dotenv==1.4.2
django-extensions==3.1.5
django-filter==21.1
django-imagekit==4.1.0
django-js-asset==2.0.0
django-location-field==2.1.0
django-model-utils==4.2.0
django-mptt==0.13.4
django-oauth-toolkit==1.7.1
django-otp==1.1.3
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-gis==1.0
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.2.1
drf-extra-fields==3.4.0
drf-reset-password==1.2.9
drf-social-oauth2==1.2.1
ecdsa==0.17.0
email-validator==1.1.3
emails==0.5.15
fastapi==0.70.1
firebase-admin==5.2.0
fonttools==4.33.3
geographiclib==1.52
geojson==2.5.0
geopy==2.2.0
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
gunicorn==20.1.0
h11==0.9.0
httplib2==0.20.4
httptools==0.1.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
incremental==21.3.0
inflect==5.4.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==2.11.3
jwcrypto==1.0
kiwisolver==1.4.2
kombu==4.6.11
leaflet==0.0.3
loguru==0.6.0
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
more-itertools==8.12.0
msgpack==1.0.4
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.2.0
openapi-codec==1.3.2
OSMPythonTools==0.3.4
p5py==1.0.0
packaging==21.3
pandas==1.4.2
passlib==1.7.4
pep517==0.12.0
pilkit==2.0
Pillow==9.0.1
pkgconfig==1.5.5
pluggy==0.13.1
premailer==3.10.0
progressbar2==4.0.0
proto-plus==1.20.6
protobuf==3.20.1
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==5.4.3
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
python-utils==3.1.0
python3-openid==3.2.0
pytz==2021.3
raven==6.10.0
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
service-identity==21.1.0
shortuuid==1.0.8
signals==0.0.2
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.34
sqlparse==0.4.2
starlette==0.16.0
swapper==1.3.0
tenacity==6.3.1
tomli==2.0.1
tqdm==4.63.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2021.5
ujson==5.2.0
Unidecode==1.3.4
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.11.8
uvloop==0.16.0
vine==1.3.0
wcwidth==0.2.5
websockets==8.1
Werkzeug==2.1.1
whitenoise==6.0.0
world==4.1.1
wrapt==1.14.0
WTForms==3.0.1
xarray==2022.3.0
zipp==3.8.0
zope.interface==5.4.0
